[Feat] DB, local 카테고리 등록, 삭제 기능 구현
DB, Local 에서 카테고리 대분류, 소분류 별 등록, 삭제 기능 구현
해당 위치 파일 업로드 기능 구현

+++ fileFinder/src/main/java/com/kmu/filefinder/main/service/CategoryService.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
+import com.kmu.filefinder.common.utils.ConvertType;
+import com.kmu.filefinder.common.utils.Verification;
 import com.kmu.filefinder.main.dao.CategoryDAO;
 import com.kmu.filefinder.main.dto.CategoryDTO;
 import com.kmu.filefinder.main.mapper.MainMapper;
@@ -30,7 +32,7 @@ public class CategoryService {
 		if (!validateFolderNameExistence(dto.getCategory_nm())) {
 			return 2;
 		}
-		if(dto.getCategory_nm().equals("")) {
+		if (dto.getCategory_nm().equals("")) {
 			return 3;
 		}
 		String urlPath = createFile(dto);
@@ -48,9 +50,9 @@ public class CategoryService {
 
 	// 로컬 디렉토리 파일 생성
 	private String createFile(CategoryDTO dto) {
-		
+
 		String urlPath = "";
-		
+
 		if (dto.getCategory_order() == 1) {
 			urlPath = path + dto.getCategory_nm();
 			File folder = new File(urlPath);
@@ -61,16 +63,57 @@ public class CategoryService {
 			File folder = new File(urlPath);
 			folder.mkdirs();
 		}
-		
+
 		return urlPath;
 	}
-	
-	public int deleteSmallCategory(int i_category) {
+
+	public int deleteSmallCategory(String id) {
+		int i_category = ConvertType.convertStringToInt(id);
+		if(Verification.checkNullByString(id)) {
+			return 2;
+		}
+		String path = getPath(i_category);
+		deleteLocalCategory(path);
+
 		return mainMapper.deleteSmallCategory(i_category);
 	}
+
+	public int deleteLargeCategory(String id) {
+		int i_category = ConvertType.convertStringToInt(id);
+		if(Verification.checkNullByString(id)) {
+			return 2;
+		}
+		String path = getPath(i_category);
+		deleteLocalCategory(path);
+
+		mainMapper.deleteLargeCategory(i_category); // 대분류 삭제
+		return mainMapper.deleteSmallCategory(i_category); // 소분류 삭제
+	}
+
 	
-	public int deleteLargeCategory(int i_category) {
-		mainMapper.deleteLargeCategory(i_category);
-		return mainMapper.deleteSmallCategory(i_category);
+
+	private String getPath(int i_category) {
+		return mainMapper.getCategoryPathByIcategory(i_category);
+	}
+
+	private void deleteLocalCategory(String path) {
+
+		File folder = new File(path);
+		try {
+			if (folder.exists()) {
+				File[] folder_list = folder.listFiles(); // 파일리스트 얻어오기
+				for (int i = 0; i < folder_list.length; i++) {
+					if (folder_list[i].isFile()) {
+						folder_list[i].delete();
+					} else { // 폴더안에 폴더가 있을 경우
+						deleteLocalCategory(folder_list[i].getPath()); // 재귀함수호출
+					}
+					folder_list[i].delete();
+				}
+				folder.delete(); // 폴더 삭제
+			}
+		} catch (Exception e) {
+			e.getStackTrace();
+		}
 	}
 }
